Redirect everything but LetsEncrypt for website_rootredirection
In 5f3edf1 we implemented 'redirect everything but LetsEncrypt' to https (from http).

We also need the same logic to redirect everything to a given destination
(if website_rootredirection is set).

So basically the following happens if 'website_httpsonly=true' and 'website_rootredirection' are set:

* redirect 'everything but LetsEncrypt' from http://$DOMAIN -> https://$DOMAIN
* (and then) redirect 'everything but LetsEncrypt' from https://$DOMAIN -> $website_rootredirection

To accomplish a direct redirection from http://$DOMAIN -> $website_rootredirection
one has to set 'website_rootredirection' and 'website_httpsonly=false'.

Moved '{% endif %}{# website_rootredirection #}' below the '<Directorymatch "^/.*/\.git+/">' section
because we only need it if 'website_rootredirection' is not set.

While at it update the 'Role Variables' section in README.md to just
link to `defaults/main.yml` so the role variable documentation doesn't
have to be updated twice.