Fix ansible-lint errors/warnings
The latest version of ansible-lint (v4.3.5) showed the following errors:

  [208] File permissions unset or incorrect
  tasks/main.yml:23
  Task/Handler: Deploy apache configuration for {{ website_domain }}

  [208] File permissions unset or incorrect
  tasks/main.yml:27
  Task/Handler: Deploy custom apache configuration for {{ website_domain }}

Added user and group owner and permissions of the destination file to make
ansible-lint happy and while at it get rid of the hard to read one-liner tasks.

Also matched consistency by putting 'not' in front instead of using 'is not
defined' in the 'when' statement.

Changed the source path of the sites-enabled symlink to be a relative path to
match the behavior of the 'a2ensite' command which also creates relative path
links.

Added white spaces around the task blocks because they are considered good
practice and make the playbook more readable.
(See: https://docs.ansible.com/ansible/latest/user_guide/playbooks_best_practices.html#playbooks-tips-and-tricks)