website.conf.j2: Remove double quotes around ProxyPass + ProxyPassReverse
In d5dbbfe8 the double quotes were added,
but it turned out that if you want to add extra options like
'nocanon' to the ProxyPass Directive the get part of the ProxyPass path.

Therefor reverting that change. If you need quotes,
set them in the website_proxy + website_proxyreverse variables.